patient_management_backend/requirements.txt
pymongo>=4.0.0
alembic==1.11.1
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.28.0
bcrypt==4.3.0
black==23.7.0
certifi==2025.7.9
cffi==1.17.1
click==8.2.1
cryptography==45.0.5
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.100.1
greenlet==3.2.3
h11==0.14.0
httpcore==0.17.3
httptools==0.6.4
httpx==0.24.1
idna==3.10
iniconfig==2.1.0
isort==5.12.0
Mako==1.3.10
MarkupSafe==3.0.2
mypy==1.4.1
mypy_extensions==1.1.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.1.1
pydantic_core==2.4.0
pytest==7.4.0
pytest-asyncio==0.21.1
python-decouple==3.8
python-dotenv==1.1.1
python-jose==3.3.0
python-multipart==0.0.20
PyYAML==6.0.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.19
starlette==0.27.0
typing_extensions==4.14.1
uvicorn==0.23.1
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
